pyproject.toml
[project]
name = "hch"
version = "0.1.0"
description = "Awesome HITSZ course hunter."
readme = "README.md"
authors = [{ name = "AbyssSkb", email = "abyssskb@outlook.com" }]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pycryptodome>=3.23.0",
    "pydantic>=2.11.7",
    "selectolax>=0.3.32",
    "typer>=0.16.0",
]

[project.scripts]
hch = "hch.main:app"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["pre-commit>=4.2.0", "pyright>=1.1.403", "ruff>=0.12.4"]
